+++ packages/service-core/src/storage/WriteCheckpointAPI.ts
@@ -50,10 +50,36 @@ export interface ManagedWriteCheckpointFilters extends BaseWriteCheckpointIdenti
   heads: Record<string, string>;
 }
 
+export interface WriteCheckpointResult {
+  /**
+   * Write checkpoint id (also referred to as client_id).
+   *
+   * If null, there is no write checkpoint for the client.
+   */
+  id: bigint | null;
+
+  /**
+   * LSN for the checkpoint.
+   *
+   * This will change when we support multiple connections.
+   *
+   * For managed write checkpoints, this LSN must be exceeded by the checkpoint / replication head to be valid.
+   *
+   * For custom write checkpoints, this will be null, and the write checkpoint is valid for all LSNs.
+   */
+  lsn: string | null;
+}
+
 export type ManagedWriteCheckpointOptions = ManagedWriteCheckpointFilters;
 
 export type SyncStorageLastWriteCheckpointFilters = BaseWriteCheckpointIdentifier | ManagedWriteCheckpointFilters;
 export type LastWriteCheckpointFilters = CustomWriteCheckpointFilters | ManagedWriteCheckpointFilters;
+
+export interface WatchUserWriteCheckpointOptions {
+  user_id: string;
+  sync_rules_id: number;
+  signal: AbortSignal;
+}
 
 export interface BaseWriteCheckpointAPI {
   readonly writeCheckpointMode: WriteCheckpointMode;
@@ -68,7 +94,6 @@ export interface BaseWriteCheckpointAPI {
  */
 export interface SyncStorageWriteCheckpointAPI extends BaseWriteCheckpointAPI {
   batchCreateCustomWriteCheckpoints(checkpoints: BatchedCustomWriteCheckpointOptions[]): Promise<void>;
-  createCustomWriteCheckpoint(checkpoint: BatchedCustomWriteCheckpointOptions): Promise<bigint>;
   lastWriteCheckpoint(filters: SyncStorageLastWriteCheckpointFilters): Promise<bigint | null>;
 }
 
@@ -78,8 +103,9 @@ export interface SyncStorageWriteCheckpointAPI extends BaseWriteCheckpointAPI {
  */
 export interface WriteCheckpointAPI extends BaseWriteCheckpointAPI {
   batchCreateCustomWriteCheckpoints(checkpoints: CustomWriteCheckpointOptions[]): Promise<void>;
-  createCustomWriteCheckpoint(checkpoint: CustomWriteCheckpointOptions): Promise<bigint>;
   lastWriteCheckpoint(filters: LastWriteCheckpointFilters): Promise<bigint | null>;
+
+  watchUserWriteCheckpoint(options: WatchUserWriteCheckpointOptions): AsyncIterable<WriteCheckpointResult>;
 }
 
 export const DEFAULT_WRITE_CHECKPOINT_MODE = WriteCheckpointMode.MANAGED;